+++ jtodo/app/views/admin/users/edit.html.erb
@@ -1,4 +1,4 @@
-<% provide(:title, 'Edit User') %>
+<% provide(:title, 'Edit user') %>
 <h1>Update Your Profile</h1>
 
 <%= render 'form', user: @user %>
