+++ jtodo/app/views/admin/users/index.html.erb
@@ -1,0 +1,25 @@
+<% provide(:title, 'All users') %>
+<h1>All users</h1>
+
+<table>
+  <thead>
+    <tr>
+      <th><%= t('users.label.name') %></th>
+      <th><%= t('users.label.is_admin') %></th>
+      <th><%= t('users.label.task_count') %></th>
+      <th><%= t('users.label.created_at') %></th>
+    </tr>
+  </thead>
+  <tbody>
+    <% @users.each do |user| %>
+      <tr>
+        <td><%= link_to user.name, edit_admin_user_path(user) %></td>
+        <td><%= user.is_admin %></td>
+        <td><%= link_to user.tasks.count, admin_user_tasks_path(user) %></td>
+        <td><%= user.created_at %></td>
+      </tr>
+    <% end %>
+  </tbody>
+</table>
+<%= paginate @users %>
+<br>
