--- jtodo/app/views/tasks/_task.html.erb
@@ -1,6 +1,0 @@
-<tr>
-  <td><%= task.status_i18n %></td>
-  <td><%= link_to task.title, task %></td>
-  <td><%= task.priority_i18n %></td>
-  <td><%= task.due_date %></td>
-</tr>
