--- jtodo/app/views/tasks/_tasks.html.erb
@@ -1,15 +1,0 @@
-<table>
-  <thead>
-    <tr>
-      <th><%= t('tasks.label.status') %></th>
-      <th><%= t('tasks.label.title') %></th>
-      <th><%= t('tasks.label.priority')%><%= link_sort 'priority' %></th>
-      <th><%= t('tasks.label.due_date')%><%= link_sort 'due_date' %></th>
-    </tr>
-  </thead>
-  <tbody>
-    <%= render @tasks %>
-  </tbody>
-</table>
-<%= paginate @tasks %>
-<br>
