--- jtodo/app/views/users/_user.html.erb
@@ -1,6 +1,0 @@
-<tr>
-  <td><%= link_to user.name, user %></td>
-  <td><%= user.is_admin %></td>
-  <td><%= user.tasks.count %></td>
-  <td><%= user.created_at %></td>
-</tr>
