--- jtodo/app/views/users/_users.html.erb
@@ -1,15 +1,0 @@
-<table>
-  <thead>
-    <tr>
-      <th><%= t('users.label.name') %></th>
-      <th><%= t('users.label.is_admin') %></th>
-      <th><%= t('users.label.task_count') %></th>
-      <th><%= t('users.label.created_at') %></th>
-    </tr>
-  </thead>
-  <tbody>
-    <%= render @users %>
-  </tbody>
-</table>
-<%= paginate @users %>
-<br>
