--- jtodo/app/views/users/index.html.erb
@@ -1,4 +1,0 @@
-<% provide(:title, 'All users') %>
-<h1>All users</h1>
-
-<%= render 'users' %>
